requirements.txt
sqlmodel~=0.0.22
python-dotenv~=1.0.1
fastapi~=0.115.5
mariadb~=1.1.11
pydantic-settings~=2.6.1
SQLAlchemy~=2.0.36